# openapi.yaml
openapi: 3.1.0
info:
  title: Qaida Places API
  version: 1.0.0
  summary: API inferred from the Qaida mobile UI kit screens.
  description: |
    This specification was derived from the HTML screens in the design folder.
    It describes the backend contract implied by the search, recommendations,
    place details, favorites, profile, collections, and notifications flows.

    Some parts are inferred from UI behavior and should be confirmed before implementation:
    - recommendation scoring logic
    - booking flows mentioned in notifications
    - collection vs default favorites behavior
    - push notification delivery strategy
servers:
  - url: https://api.qaida.app/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered and tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Refreshed tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user and revoke current tokens
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Tokens revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /user/notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      parameters:
        - in: query
          name: type
          schema:
            type: array
            items:
              $ref: '#/components/schemas/NotificationType'
          style: form
          explode: false
        - in: query
          name: read
          schema:
            type: boolean
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/notifications/{notificationId}:
    patch:
      tags: [Notifications]
      summary: Update notification state
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationRequest'
      responses:
        '200':
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Notifications]
      summary: Delete notification
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '204':
          description: Notification deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /user/notification-preferences:
    get:
      tags: [Notifications]
      summary: Get notification preferences
      responses:
        '200':
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Notifications]
      summary: Save notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferences'
      responses:
        '200':
          description: Notification preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/notification-preferences/subscribe:
    post:
      tags: [Notifications]
      summary: Enable notification subscription for the current user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceToken:
                  type: string
                platform:
                  type: string
                  enum: [ios, android, web]
      responses:
        '200':
          description: Notifications enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    NotificationId:
      in: path
      name: notificationId
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    BudgetRange:
      type: string
      enum: ['$', '$$', '$$$']
    CompanyType:
      type: string
      enum: [solo, couple, friends, family]
...
    NotificationType:
      type: string
      enum: [new_place, reminder, saved_place, recommendation, booking]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: uri
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        notificationPreferences:
          $ref: '#/components/schemas/NotificationPreferences'
      required: [id, name, email, preferences, notificationPreferences]
    UserPreferences:
      type: object
      properties:
        companyType:
          $ref: '#/components/schemas/CompanyType'
        favoriteCategories:
          type: array
          items:
            type: string
        budget:
          $ref: '#/components/schemas/BudgetRange'
      additionalProperties: false
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        message:
          type: string
        actionUrl:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, type, title, message, read, createdAt]
    NotificationPreferences:
      type: object
      properties:
        enabled:
          type: boolean
        enabledTypes:
          type: array
          items:
            $ref: '#/components/schemas/NotificationType'
      required: [enabled, enabledTypes]
...
    UpdateNotificationRequest:
      type: object
      properties:
        read:
          type: boolean
      required: [read]
...
    AcceptedResponse:
      type: object
      properties:
        accepted:
          type: boolean
          const: true
      required: [accepted]
...
    NotificationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        unreadCount:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
      required: [items, unreadCount, total]
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        avatarUrl:
          type: string
          format: uri
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
      required: [email, password]
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
      required: [id, name, email]
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: bearer
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/AuthUser'
      required: [accessToken, refreshToken, tokenType, expiresIn, user]